consolidate store
